overhaul to ES6 modules
