+++ tools/codegen/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Code Generator. Converts Catena JSON device models into source code representations in various languages.",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
